corrected vertical 3 and 4 and multiaxis
JP is a genius and thinks that it's the code's fault, so I'm fixing the
engineering mistake. The mistake was that vertical motors 3 and 4 are
now counter rotating and were not before, so I had to flip the direction
of every instance these two motors show up.

Also fixed multiaxis, directions were backwards. This was an ACTUAL code
issue.